Put the capitalize and trim classes in seperate files
