+++ capitalize.rb
@@ -1,0 +1,7 @@
+require "./decorator.rb"
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    "#{@nameable.correct_name}.capitalize()"
+  end
+end
